+++ src/server/visibility.rs
@@ -1,5 +1,8 @@
 pub mod client_visibility;
+pub mod filters_mask;
 pub mod registry;
+
+use core::marker::PhantomData;
 
 use bevy::{
     ecs::{component::Immutable, entity_disabling::Disabled},
@@ -7,18 +10,24 @@ use bevy::{
 };
 use log::debug;
 
+use crate::shared::replication::registry::{
+    ReplicationRegistry, command_fns::MutWrite, component_mask::ComponentMask,
+};
 use client_visibility::ClientVisibility;
-use registry::FilterRegistry;
+use registry::{FilterRegistry, VisibilityScope};
 
 /// Remote visibility functions for [`App`].
 pub trait AppVisibilityExt {
     /**
-    Registers a component as a remote visibility filter for entities.
+    Registers a component as a remote visibility filter.
 
-    An entity will be visible to a client if it has all the filter components
-    present on the entity, and [`VisibilityFilter::is_visible`] returns `true` for each of them.
+    This component needs to be inserted on both the client entity and replicated entities.
+    If [`VisibilityFilter::is_visible`] on the client;s component returns `false` for the
+    corresponding component on a replicated entity, the associated [`VisibilityFilter::Scope`]
+    (entity or components) becomes hidden for the client.
 
-    To check whether an entity is hidden from a client based on all filters, use [`ClientVisibility::is_hidden`].
+    If the data was previously visible, it will be despawned or removed. If the component is
+    missing on either the client or a replicated entity, it is treated as evaluating to `false`.
 
     To keep the representation compact, the total number of registered filters cannot exceed [`u32::MAX`].
     But a filter can itself represent multiple flags using a bitmask. See the example in [`VisibilityFilter`].
@@ -59,8 +68,10 @@ pub trait AppVisibilityExt {
     }
 
     impl VisibilityFilter for Guild {
+        type Scope = Entity;
+
         fn is_visible(&self, entity_filter: &Self) -> bool {
-        self == entity_filter
+            self == entity_filter
         }
     }
     ```
@@ -73,8 +84,13 @@ impl AppVisibilityExt for App {
         debug!("adding visibility filter `{}`", ShortName::of::<F>());
 
         self.world_mut()
-            .resource_mut::<FilterRegistry>()
-            .register::<F>();
+            .resource_scope(|world, mut registry: Mut<FilterRegistry>| {
+                world.resource_scope(
+                    |world, mut replication_registry: Mut<ReplicationRegistry>| {
+                        registry.register::<F>(world, &mut replication_registry);
+                    },
+                )
+            });
 
         self.add_observer(hide_for_new_clients::<F>)
             .add_observer(on_insert::<F>)
@@ -89,14 +105,14 @@ fn hide_for_new_clients<F: VisibilityFilter>(
     entities: Query<Entity, With<F>>,
 ) {
     if let Ok(mut visibility) = clients.get_mut(insert.entity) {
-        let filter_bit = registry.get::<F>();
+        let bit = registry.bit::<F>();
         for entity in &entities {
             debug!(
                 "hiding `{entity}` from client `{}` without `{}` filter",
                 insert.entity,
                 ShortName::of::<F>(),
             );
-            visibility.set_visibility(entity, filter_bit, false);
+            visibility.set(entity, bit, false);
         }
     }
 }
@@ -107,7 +123,7 @@ fn on_insert<F: VisibilityFilter>(
     entities: Query<(Entity, &F), (Without<ClientVisibility>, Allow<Disabled>)>,
     mut clients: Query<(Entity, Option<&F>, &mut ClientVisibility)>,
 ) {
-    let filter_bit = registry.get::<F>();
+    let bit = registry.bit::<F>();
     if let Ok((client_entity, client_component, mut visibility)) = clients.get_mut(insert.entity) {
         let client_component = client_component.unwrap();
         for (entity, component) in &entities {
@@ -116,7 +132,7 @@ fn on_insert<F: VisibilityFilter>(
                 "updating `{}` filter on client `{client_entity}` to `{visible}` for `{entity}`",
                 ShortName::of::<F>(),
             );
-            visibility.set_visibility(entity, filter_bit, visible);
+            visibility.set(entity, bit, visible);
         }
     } else {
         let (entity, component) = entities.get(insert.entity).unwrap();
@@ -126,7 +142,7 @@ fn on_insert<F: VisibilityFilter>(
                 "updating `{}` filter on `{entity}` to `{visible}` for client `{client_entity}`",
                 ShortName::of::<F>(),
             );
-            visibility.set_visibility(insert.entity, filter_bit, visible);
+            visibility.set(insert.entity, bit, visible);
         }
     }
 }
@@ -137,7 +153,7 @@ fn on_remove<F: VisibilityFilter>(
     mut clients: Query<&mut ClientVisibility>,
     entities: Query<Entity, (With<F>, Without<ClientVisibility>)>,
 ) {
-    let filter_bit = registry.get::<F>();
+    let bit = registry.bit::<F>();
     if let Ok(mut visibility) = clients.get_mut(remove.entity) {
         for entity in &entities {
             debug!(
@@ -145,7 +161,7 @@ fn on_remove<F: VisibilityFilter>(
                 remove.entity,
                 ShortName::of::<F>(),
             );
-            visibility.set_visibility(entity, filter_bit, false);
+            visibility.set(entity, bit, false);
         }
     } else {
         debug!(
@@ -154,7 +170,7 @@ fn on_remove<F: VisibilityFilter>(
             remove.entity
         );
         for mut visibility in &mut clients {
-            visibility.set_visibility(remove.entity, filter_bit, true);
+            visibility.set(remove.entity, bit, true);
         }
     }
 }
@@ -171,11 +187,14 @@ Visible if the filter is present on both the entity and the client:
 ```
 # use bevy::prelude::*;
 # use bevy_replicon::prelude::*;
+/// Only ghost players can see ghosts.
 #[derive(Component)]
 #[component(immutable)] // Component should be immutable.
-struct RemoteVisible;
+struct Ghost;
 
-impl VisibilityFilter for RemoteVisible {
+impl VisibilityFilter for Ghost {
+    type Scope = Entity;
+
     fn is_visible(&self, _entity_filter: &Self) -> bool {
         true
     }
@@ -192,8 +211,10 @@ Visible if the entity and the client belong to the same team:
 struct Team(u8);
 
 impl VisibilityFilter for Team {
+    type Scope = Entity;
+
     fn is_visible(&self, entity_filter: &Self) -> bool {
-    self == entity_filter
+        self == entity_filter
     }
 }
 ```
@@ -217,16 +238,139 @@ bitflags! {
 }
 
 impl VisibilityFilter for RemoteVisibility {
+    type Scope = Entity;
+
     fn is_visible(&self, entity_filter: &Self) -> bool {
-    entity_filter.contains(*self)
+        entity_filter.contains(*self)
     }
 }
 ```
 */
 pub trait VisibilityFilter: Component<Mutability = Immutable> {
-    /// Returns `true` if a client with this component should see an entity with this component.
+    /// Defines what data is affected when the filter denies visibility.
+    ///
+    /// - To hide the entire entity, this type must be [`Entity`].
+    /// - To hide a single component on the entity, this type must be [`ComponentScope`].
+    /// - To hide more than one component on the entity, this type must be a tuple of those [`Component`]s.
+    ///
+    /// # Examples
+    ///
+    /// Hide the entire entity:
+    ///
+    /// ```
+    /// # use bevy::prelude::*;
+    /// # use bevy_replicon::prelude::*;
+    /// #[derive(Component, PartialEq)]
+    /// #[component(immutable)]
+    /// struct Team(u8);
+    ///
+    /// impl VisibilityFilter for Team {
+    ///     type Scope = Entity;
+    ///
+    ///     fn is_visible(&self, entity_filter: &Self) -> bool {
+    ///         self == entity_filter
+    ///     }
+    /// }
+    /// ```
+    ///
+    /// Hide only a single component:
+    ///
+    /// ```
+    /// # use bevy::prelude::*;
+    /// # use bevy_replicon::prelude::*;
+    /// #[derive(Component, PartialEq)]
+    /// #[component(immutable)]
+    /// struct Team(u8);
+    ///
+    /// impl VisibilityFilter for Team {
+    ///     type Scope = ComponentScope<Health>;
+    ///
+    ///     fn is_visible(&self, entity_filter: &Self) -> bool {
+    ///         self == entity_filter
+    ///     }
+    /// }
+    ///
+    /// #[derive(Component)]
+    /// struct Health(u8);
+    /// ```
+    ///
+    /// Hide multiple components:
+    ///
+    /// ```
+    /// # use bevy::prelude::*;
+    /// # use bevy_replicon::prelude::*;
+    /// #[derive(Component, PartialEq)]
+    /// #[component(immutable)]
+    /// struct Team(u8);
+    ///
+    /// impl VisibilityFilter for Team {
+    ///     type Scope = (Health, Stats);
+    ///
+    ///     fn is_visible(&self, entity_filter: &Self) -> bool {
+    ///         self == entity_filter
+    ///     }
+    /// }
+    ///
+    /// #[derive(Component)]
+    /// struct Health(u8);
+    ///
+    /// #[derive(Component)]
+    /// struct Stats {
+    /// // ...
+    /// }
+    /// ```
+    type Scope: FilterScope;
+
+    /// Returns `true` if a client with this component should see [`Self::Scope`] for an entity with this component.
     fn is_visible(&self, entity_filter: &Self) -> bool;
 }
+
+/// Associates the type with a visibility scope.
+pub trait FilterScope {
+    /// Returns data that should be hidden when [`VisibilityFilter::is_visible`] returns `false`.
+    fn visibility_scope(world: &mut World, registry: &mut ReplicationRegistry) -> VisibilityScope;
+}
+
+/// A scope for a single component `A`.
+///
+/// We can't implement [`FilterScope`] for both tuples and all types that implement [`Component`].
+/// This is why this wrapper is needed to set the scope for only a single component.
+pub struct ComponentScope<A: Component>(PhantomData<A>);
+
+impl<C: Component<Mutability: MutWrite<C>>> FilterScope for ComponentScope<C> {
+    fn visibility_scope(world: &mut World, registry: &mut ReplicationRegistry) -> VisibilityScope {
+        let mut mask = ComponentMask::default();
+        let (index, _) = registry.init_component_fns::<C>(world);
+        mask.insert(index);
+        VisibilityScope::Components(mask)
+    }
+}
+
+impl FilterScope for Entity {
+    fn visibility_scope(
+        _world: &mut World,
+        _registry: &mut ReplicationRegistry,
+    ) -> VisibilityScope {
+        VisibilityScope::Entity
+    }
+}
+
+macro_rules! impl_filter_scope {
+    ($($C:ident),*) => {
+        impl<$($C: Component<Mutability: MutWrite<$C>>),*> FilterScope for ($($C,)*) {
+            fn visibility_scope(world: &mut World, registry: &mut ReplicationRegistry) -> VisibilityScope {
+                let mut mask = ComponentMask::default();
+                $(
+                    let (index, _) = registry.init_component_fns::<$C>(world);
+                    mask.insert(index);
+                )*
+                VisibilityScope::Components(mask)
+            }
+        }
+    };
+}
+
+variadics_please::all_tuples!(impl_filter_scope, 2, 10, C);
 
 #[cfg(test)]
 mod tests {
@@ -238,53 +382,60 @@ mod tests {
     fn after_clients() {
         let mut app = App::new();
         app.init_resource::<FilterRegistry>()
+            .init_resource::<ReplicationRegistry>()
             .add_visibility_filter::<A>();
 
         let client1 = app.world_mut().spawn((ClientVisibility::default(), A)).id();
         let client2 = app.world_mut().spawn(ClientVisibility::default()).id();
         let entity = app.world_mut().spawn(A).id();
 
+        let registry = app.world().resource::<FilterRegistry>();
         let visibility1 = app.world().get::<ClientVisibility>(client1).unwrap();
-        assert!(!visibility1.is_hidden(entity));
+        assert!(!visibility1.get(entity).is_hidden(registry));
 
         let visibility2 = app.world().get::<ClientVisibility>(client2).unwrap();
-        assert!(visibility2.is_hidden(entity));
+        assert!(visibility2.get(entity).is_hidden(registry));
     }
 
     #[test]
     fn before_clients() {
         let mut app = App::new();
         app.init_resource::<FilterRegistry>()
+            .init_resource::<ReplicationRegistry>()
             .add_visibility_filter::<A>();
 
         let entity = app.world_mut().spawn(A).id();
         let client1 = app.world_mut().spawn((ClientVisibility::default(), A)).id();
         let client2 = app.world_mut().spawn(ClientVisibility::default()).id();
 
+        let registry = app.world().resource::<FilterRegistry>();
         let visibility1 = app.world().get::<ClientVisibility>(client1).unwrap();
-        assert!(!visibility1.is_hidden(entity));
+        assert!(!visibility1.get(entity).is_hidden(registry));
 
         let visibility2 = app.world().get::<ClientVisibility>(client2).unwrap();
-        assert!(visibility2.is_hidden(entity));
+        assert!(visibility2.get(entity).is_hidden(registry));
     }
 
     #[test]
     fn remove_filter_from_entity() {
         let mut app = App::new();
         app.init_resource::<FilterRegistry>()
+            .init_resource::<ReplicationRegistry>()
             .add_visibility_filter::<A>();
 
         let client = app.world_mut().spawn(ClientVisibility::default()).id();
         let entity = app.world_mut().spawn(A).remove::<A>().id();
 
+        let registry = app.world().resource::<FilterRegistry>();
         let visibility = app.world().get::<ClientVisibility>(client).unwrap();
-        assert!(!visibility.is_hidden(entity));
+        assert!(!visibility.get(entity).is_hidden(registry));
     }
 
     #[test]
     fn remove_filter_from_client() {
         let mut app = App::new();
         app.init_resource::<FilterRegistry>()
+            .init_resource::<ReplicationRegistry>()
             .add_visibility_filter::<A>();
 
         let entity = app.world_mut().spawn(A).id();
@@ -294,14 +445,16 @@ mod tests {
             .remove::<A>()
             .id();
 
+        let registry = app.world().resource::<FilterRegistry>();
         let visibility = app.world().get::<ClientVisibility>(client).unwrap();
-        assert!(visibility.is_hidden(entity));
+        assert!(visibility.get(entity).is_hidden(registry));
     }
 
     #[test]
     fn multiple_filters() {
         let mut app = App::new();
         app.init_resource::<FilterRegistry>()
+            .init_resource::<ReplicationRegistry>()
             .add_visibility_filter::<A>()
             .add_visibility_filter::<B>();
 
@@ -312,26 +465,29 @@ mod tests {
         let client2 = app.world_mut().spawn((ClientVisibility::default(), A)).id();
         let entity = app.world_mut().spawn((A, B)).id();
 
+        let registry = app.world().resource::<FilterRegistry>();
         let visibility1 = app.world().get::<ClientVisibility>(client1).unwrap();
-        assert!(!visibility1.is_hidden(entity));
+        assert!(!visibility1.get(entity).is_hidden(registry));
 
         let visibility2 = app.world().get::<ClientVisibility>(client2).unwrap();
-        assert!(visibility2.is_hidden(entity));
+        assert!(visibility2.get(entity).is_hidden(registry));
 
         // Hide entity from the first client too.
         app.world_mut().entity_mut(client1).remove::<B>();
 
+        let registry = app.world().resource::<FilterRegistry>();
         let visibility1 = app.world().get::<ClientVisibility>(client1).unwrap();
-        assert!(visibility1.is_hidden(entity));
+        assert!(visibility1.get(entity).is_hidden(registry));
 
         // Relax visibility constraints to make it visible to both.
         app.world_mut().entity_mut(entity).remove::<B>();
 
+        let registry = app.world().resource::<FilterRegistry>();
         let visibility1 = app.world().get::<ClientVisibility>(client1).unwrap();
-        assert!(!visibility1.is_hidden(entity));
+        assert!(!visibility1.get(entity).is_hidden(registry));
 
         let visibility2 = app.world().get::<ClientVisibility>(client2).unwrap();
-        assert!(!visibility2.is_hidden(entity));
+        assert!(!visibility2.get(entity).is_hidden(registry));
     }
 
     #[derive(Component)]
@@ -339,6 +495,8 @@ mod tests {
     struct A;
 
     impl VisibilityFilter for A {
+        type Scope = Entity;
+
         fn is_visible(&self, _entity_filter: &Self) -> bool {
             true
         }
@@ -349,6 +507,8 @@ mod tests {
     struct B;
 
     impl VisibilityFilter for B {
+        type Scope = Entity;
+
         fn is_visible(&self, _entity_filter: &Self) -> bool {
             true
         }